Tree Order System Update - Almost Finished - Testable
Tree will be spawned (but does not yet auto destroy)
Tree will create and display its own thought bubble with the correct decorations in the order provided in the OrderManager when the order was created.
Order icons will change to a greyscale with green tick when correct decoration is placed.
Can be tested with Star decoration so far, others should also work but need their corresponding decoration classes to be created.

TODO:
 - Create decoration classes for bow and bauble
 - Test bow and bauble work with tree and correctly show as placed
 - Destroy tree when OOB and report points to the order manager
 - Spawn next tree in order
 - Show placed decorations on the tree somehow