Implemented Star Interactions - see description
Get gold from sack
Gold goes into Pot - instant add
Pot of gold goes onto red station (melter) - takes 15s to cook
Pot of melted gold will go onto new station (todo) with star mould to create final star

Also created a decoration independent version of the StationTriggerArea script - the current one was tailored for bauble only. Now can be set in inspector for which deco type you want the station to work on. Could be easily expanded to a List of Deco Types if we need the station to work with multiple decos. This required an update to iDecoration and as such I had to update the RawBaubleHandler to match the new interface.